🧹 Do code refactoring
