+++ src/controllers/new-members.js
@@ -1,5 +1,9 @@
+import {
+    scheduleCaptchaInvalidation
+} from '../services/handlers/new-members/captcha.js'
 import { texts } from '../static/texts.js'
 import { buttons } from '../static/buttons.js'
+import { restrictParticipant } from '../services/handlers/restrict.js'
 
 
 // FIXME: Move logic to service
@@ -9,26 +13,24 @@ async function handleNewMembers(ctx) {
     if (isInvited) {
         return
     }
-    try {
-        await ctx.telegram.restrictChatMember(
-            ctx.chat.id, ctx.from.id,
-            { can_send_messages: false }
-        )
-    } catch {
+    const { error } = await restrictParticipant('mute', {
+        api: ctx.telegram,
+        chatId: ctx.chat.id,
+        userId: ctx.from.id
+    })
+    if (error) {
         return
     }
     const greeting = await ctx.text(
         texts.other.greeting(ctx.from.id, ctx.from.first_name),
         buttons.captcha(ctx.from.id)
     )
-    setTimeout(async () => {
-        try {
-            await ctx.telegram.deleteMessage(ctx.chat.id, greeting.message_id)
-            await ctx.telegram.kickChatMember(ctx.chat.id, newMembers[0].id)
-        } catch {
-            // Never mind if I can't delete it
-        }
-    }, 30 * 1000)
+    await scheduleCaptchaInvalidation({
+        api: ctx.telegram,
+        chatId: ctx.chat.id,
+        greetingId: greeting.message_id,
+        userId: newMembers[0].id
+    })
 }
